Add final refactors
